Move all avis moderation routes into avis-routes
