+++ backoffice/src/components/backoffice/avisService.js
@@ -1,5 +1,5 @@
+import { _delete, _get, _put } from '../../utils/http-client';
 import queryString from 'query-string';
-import { _get } from '../../utils/http-client';
 import { getToken } from '../../utils/session';
 
 export const searchAvis = (options = {}) => {
@@ -16,3 +16,40 @@ export const getExportAvisUrl = (options = {}) => {
 export const getStats = () => {
     return _get(`/backoffice/avis/stats`);
 };
+
+export const maskPseudo = (id, mask) => {
+    return _put(`/backoffice/avis/${id}/pseudo`, { mask });
+};
+
+export const maskTitle = (id, mask) => {
+    return _put(`/backoffice/avis/${id}/title`, { mask });
+};
+
+export const rejectAvis = (id, reason) => {
+    return _put(`/backoffice/avis/${id}/reject`, { reason });
+};
+
+export const publishAvis = (id, qualification) => {
+    return _put(`/backoffice/avis/${id}/publish`, { qualification });
+};
+
+export const publishReponse = id => {
+    return _put(`/backoffice/avis/${id}/publishReponse`);
+};
+
+export const rejectReponse = id => {
+    return _put(`/backoffice/avis/${id}/rejectReponse`);
+};
+
+export const editAvis = (id, text) => {
+    return _put(`/backoffice/avis/${id}/edit`, { text });
+};
+
+export const resendEmail = id => {
+    return _put(`/backoffice/avis/${id}/resendEmail`);
+};
+
+export const deleteAvis = id => {
+    return _delete(`/backoffice/avis/${id}`);
+};
+
