+++ backoffice/src/components/backoffice/common/avis/Titre.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { maskTitle } from '../../moderateur/moderation-avis/moderationService';
+import { maskTitle } from '../../avisService';
 import './Titre.scss';
 
 export default class Titre extends React.Component {
