+++ backoffice/src/components/backoffice/common/avis/buttons/RejectButton.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { rejectAvis } from '../../../moderateur/moderation-avis/moderationService';
+import { rejectAvis } from '../../../avisService';
 import Modal from '../../Modal';
 import Button from '../../Button';
 import { Dropdown, DropdownDivider, DropdownItem } from '../../Dropdown';
