+++ backoffice/src/components/backoffice/common/organisme/Organisme.scss
@@ -1,4 +1,4 @@
-@import '../../../../../styles/variables';
+@import '../../../../styles/variables';
 
 .Organisme {
