+++ backoffice/src/components/backoffice/common/organisme/buttons/EditButton.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { removeEditedCourriel, resendEmailAccount } from '../gestionOrganismesService';
-import Button from '../../../common/Button';
-import { Dropdown, DropdownDivider, DropdownItem } from '../../../common/Dropdown';
+import { removeEditedCourriel, resendEmailAccount } from '../../../moderateur/gestion-organismes/gestionOrganismesService';
+import Button from '../../Button';
+import { Dropdown, DropdownDivider, DropdownItem } from '../../Dropdown';
 
 export default class EditButton extends React.Component {
